--- src/main/java/com/altenheim/kalender/interfaces/IDateFormatController.java
@@ -1,11 +1,0 @@
-package com.altenheim.kalender.interfaces;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-
-public interface IDateFormatController {
-    String formatDate(LocalDate date);
-    String formatTime(LocalTime time);
-    String formatDateTime(LocalDateTime dateTime);
-}
